+++ src/main/java/com/fmi/comet/repository/UserRepository.java
@@ -1,4 +1,3 @@
-// UserRepository.java
 package com.fmi.comet.repository;
 
 import com.fmi.comet.model.User;
@@ -27,7 +26,7 @@ public class UserRepository {
         @Override
         public User mapRow(ResultSet rs, int rowNum) throws SQLException {
             User user = new User();
-            user.setId(rs.getInt("id"));
+            user.setId(rs.getLong("id"));  // Changed to Long
             user.setUsername(rs.getString("username"));
             user.setPassword(rs.getString("password"));
             user.setRole(User.Role.valueOf(rs.getString("role")));
@@ -52,37 +51,37 @@ public class UserRepository {
     }
 
     // Find a user by ID
-    public User findUserById(Integer id) {
+    public User findUserById(Long id) {  // Changed to Long
         String sql = "SELECT * FROM users WHERE id = ? AND is_deleted = false";  // Ensure you're retrieving only non-deleted users
         return jdbcTemplate.queryForObject(sql, USER_ROW_MAPPER, id);
     }
 
     // Mark a user as deleted (soft delete)
-    public void markUserAsDeleted(Integer id) {
+    public void markUserAsDeleted(Long id) {  // Changed to Long
         String sql = "UPDATE users SET is_deleted = true, deleted_at = ? WHERE id = ?";
         jdbcTemplate.update(sql, LocalDateTime.now(), id);
     }
 
     // Update user role
-    public void updateUserRole(Integer id, User.Role role) {
+    public void updateUserRole(Long id, User.Role role) {  // Changed to Long
         String sql = "UPDATE users SET role = ? WHERE id = ?";
         jdbcTemplate.update(sql, role.name(), id);
     }
 
     // Add a friend
-    public void addFriend(Integer userId, Integer friendId) {
+    public void addFriend(Long userId, Long friendId) {  // Changed to Long
         String sql = "INSERT INTO user_friends (user_id, friend_id) VALUES (?, ?)";
         jdbcTemplate.update(sql, userId, friendId);
     }
 
     // Remove a friend
-    public void removeFriend(Integer userId, Integer friendId) {
+    public void removeFriend(Long userId, Long friendId) {  // Changed to Long
         String sql = "DELETE FROM user_friends WHERE user_id = ? AND friend_id = ?";
         jdbcTemplate.update(sql, userId, friendId);
     }
 
     // Get all friends for a user
-    public List<User> findFriends(Integer userId) {
+    public List<User> findFriends(Long userId) {  // Changed to Long
         String sql = """
             SELECT u.* FROM users u
             INNER JOIN user_friends uf ON u.id = uf.friend_id
