+++ src/main/java/com/fmi/comet/service/UserService.java
@@ -39,29 +39,29 @@ public class UserService {
     }
 
     // Soft delete a user
-    public void softDeleteUser(Integer id) {
-        userRepository.markUserAsDeleted(id);
+    public void softDeleteUser(Long id) {  // Changed to Long
+        userRepository.markUserAsDeleted(id);  // Changed to Long
     }
 
     // Update user role
-    public User updateRole(Integer id, User.Role role) {
-        userRepository.updateUserRole(id, role);
-        return userRepository.findUserById(id);
+    public User updateRole(Long id, User.Role role) {  // Changed to Long
+        userRepository.updateUserRole(id, role);  // Changed to Long
+        return userRepository.findUserById(id);  // Changed to Long
     }
 
     // Add a friend to a user
-    public void addFriend(Integer userId, Integer friendId) {
-        userRepository.addFriend(userId, friendId);
+    public void addFriend(Long userId, Long friendId) {  // Changed to Long
+        userRepository.addFriend(userId, friendId);  // Changed to Long
     }
 
     // Remove a friend from a user
-    public void removeFriend(Integer userId, Integer friendId) {
-        userRepository.removeFriend(userId, friendId);
+    public void removeFriend(Long userId, Long friendId) {  // Changed to Long
+        userRepository.removeFriend(userId, friendId);  // Changed to Long
     }
 
     // Get all friends for a user
-    public List<User> getFriends(Integer userId) {
-        return userRepository.findFriends(userId);
+    public List<User> getFriends(Long userId) {  // Changed to Long
+        return userRepository.findFriends(userId);  // Changed to Long
     }
 
     // Register a new user (with password encoding)
@@ -73,6 +73,6 @@ public class UserService {
         userRepository.insertUser(user);  // Use insertUser() instead of save()
 
         // Return the user with the updated ID (after insert)
-        return userRepository.findUserById(user.getId());
+        return userRepository.findUserById(user.getId());  // Changed to Long
     }
 }
